sample2/api/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
celery
django_celery_results
django_celery_beat
ptvsd
tzdata
selenium
webdriver_manager